requirements.txt
argcomplete==1.10.0
argon2-cffi==20.1.0
asgiref==3.3.1
attrs==20.1.0
backcall==0.2.0
beautifulsoup4==4.8.0
bleach==3.1.5
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dictobj==0.4
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
et-xmlfile==1.0.1
extract-msg==0.23.1
filemagic==1.6
Flask==1.1.2
h11==0.11.0
h2==4.0.0
hpack==4.0.0
htmldate==0.7.0
hyperframe==6.0.0
idna==2.10
IMAPClient==2.1.0
install==1.3.3
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jstree==0.5
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
justext @ http://corpus.tools/raw-attachment/wiki/Downloads/justext-3.0.tar.gz
kaitaistruct==0.9
kiwisolver==1.3.1
langdetect==1.0.8
langid==1.1.6
ldap3==2.8.1
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mitmproxy==6.0.2
msgpack==1.0.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.1
olefile==0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
passlib==1.7.4
pdf2image==1.14.0
pdfminer==20191125
pdfminer.six==20181108
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.6.1
pygtrie==2.4.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-docx==0.8.10
python-magic==0.4.18
python-pptx==0.6.18
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
readability-lxml==0.8.1
regex==2021.4.4
requests==2.24.0
requests-file==1.5.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Send2Trash==1.5.0
sentence-splitter==1.4
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
surt==0.3.1
terminado==0.8.3
tesserocr==2.5.1
testpath==0.4.4
textract==1.6.3
tldextract==2.2.3
torender==0.1.0
tornado==6.0.4
trafilatura==0.5.1
traitlets==4.3.3
tzlocal==1.5.1
urllib3==1.25.10
urwid==2.1.2
Warcat==2.2.5
warcio==1.7.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wsproto==1.0.0
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
zstandard==0.14.1
